sdk agent/requirements.txt
# Core Dependencies
anthropic>=0.40.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiohttp>=3.9.0
websocket-client>=1.6.0
signalrcore>=0.9.5  # For TopStepX WebSocket streaming

# Discord Integration
discord.py>=2.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Technical Indicators
# ta-lib>=0.4.0  # Optional: Requires system install first (brew install ta-lib)
# For now, we'll use pandas built-in functions

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
requests>=2.31.0
pytz>=2023.3
